Improve `Decoder` API
The decoders have to endure complicated tracking of bytes read when
delegating to sub-decoders, repeating the same code. So this changes the
decoder to require a different method called `decode_chunk` which takes
a mutable reference to a slice which is supposed to be updated. This
also drops the requirement to return the consumed size since that is
implied by updating the slice.

However top-level consumers prefer the previous `bytes_received` method
which simply returns the bytes. Thus the method is left in calling into
`decode_chunk` and computing how many bytes were consumed.

Unrelated: this also improves the `mapped_decoder` macro as it
previously wasn't resembling its output. The changes got tangled with
these and I don't care to disentangle them.