Add helper `Decoder` methods
Most decoders implement `Default` which allows resetting them whenever
they finish decoding. This change adds a method to do that conveniently
in as single step. This solves borrowing issues in sub-decoders and
allows decoding multiple items easilly.

Further, when calling into sub-decoders an identical code was repeating:
returning on error or when bytes were consumed and ending decoding if
not. To help making the code simpler this also adds `sub_decode` method
which returns double `Result` and thus the `?` operator can be used on
it. This condenses all those operations into a single method call
followed by `?` at the cost of wrapping the code in a helper
function/closure. A `try` block can be used in the future.