Add fn ptr helper type definitions
When `Then` and `ThenTry` combinators are used wrapped in newtypes using
a function pointer is required (until `type = impl Trait is available).
It is annoying to type the types twice in that case and synchronize them
when they change.

To make it less annoying we add simple `type` definitions for these
types that reduce the complexity.