Add `KnownMinLenDecoder`
It turns out that the `BufRead` trait is quite unpopular despite its
usefulness. Some libraries authors didn't realize that trait should be
used instead of `Read` and breaking the API is not wanted. That'd imply
`push_decode` being unusable in such scenarios because the desing of
`Decoder` required buffered reading.

To fix this we provide an additional trait that can further restrict
`Decoder`s to also provide a method for getting the number of bytes
guaranteed to be consumed. This avoids overreading making it possible to
use the (slower) `Read` trait. All standard decoders in this crate so
far are able to implement the trait, so they do implement it.